fix generic model import
